feat: remove LinkedIn in the all the names
Due to IP violation of LinkedIn. I thought the ltter of the icon is only the problem but, it seems like 'LinkedIn' in the title could be a problem. it seems to be fair to use though, they don't look like they are going to review my extension for republishing again, so, I removed all the linkedin name and will republish it with the name 'Job Scanner'.